+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlobStore.hs
@@ -64,6 +64,7 @@ module Concordium.GlobalState.Persistent.BlobStore (
     storeUpdateRef,
     loadRef,
     DirectBlobStorable (..),
+    DirectBlobHashable (..),
 
     -- * Nullable
     Nullable (..),
@@ -172,6 +173,7 @@ import Concordium.Wasm
 
 import qualified Concordium.Crypto.SHA256 as H
 import qualified Concordium.GlobalState.AccountMap.LMDB as LMDBAccountMap
+import Concordium.GlobalState.AccountMap.ModuleMap (MonadModuleMapStore)
 import Concordium.Types.HashableTo
 
 -- | A @BlobRef a@ represents an offset on a file, at
@@ -603,6 +605,12 @@ deriving via
         (MonadIO m, MonadLogger m, LMDBAccountMap.HasDatabaseHandlers r) =>
         LMDBAccountMap.MonadAccountMapStore (BlobStoreT r m)
 
+deriving via
+    (LMDBAccountMap.AccountMapStoreMonad (BlobStoreT r m))
+    instance
+        (MonadIO m, MonadLogger m, LMDBAccountMap.HasDatabaseHandlers r) =>
+        MonadModuleMapStore (BlobStoreT r m)
+
 -- | Apply a given function to modify the context of a 'BlobStoreT' operation.
 alterBlobStoreT :: (r1 -> r2) -> BlobStoreT r2 m a -> BlobStoreT r1 m a
 alterBlobStoreT f (BlobStoreT a) = BlobStoreT (a . f)
@@ -873,6 +881,23 @@ instance {-# OVERLAPPABLE #-} (MonadBlobStore m, BlobStorable m a) => DirectBlob
     loadDirect = loadRef
     {-# INLINE storeUpdateDirect #-}
     {-# INLINE loadDirect #-}
+
+-- | The @DirectBlobHashable m h a@ class defines an operation for directly loading the hash
+--  of a value of type @a@ from the blob store. This allows for a more efficient implementation
+--  than loading the full value and then computing the hash in cases where loading may be expensive.
+class (MonadBlobStore m) => DirectBlobHashable m h a where
+    -- | Load the hash of a value of type @a@ from the underlying storage.
+    --
+    -- prop> loadHash = getHashM <=< loadDirect
+    loadHash :: BlobRef a -> m h
+
+instance
+    {-# OVERLAPPABLE #-}
+    (DirectBlobStorable m a, MHashableTo m h a) =>
+    DirectBlobHashable m h a
+    where
+    loadHash = getHashM <=< loadDirect
+    {-# INLINE loadHash #-}
 
 instance (BlobStorable m a, BlobStorable m b) => BlobStorable m (a, b) where
     storeUpdate (a, b) = do